pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cta_framework"
version = "0.1.0"
description = "Event-driven CTA backtesting framework for crypto markets"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "matplotlib>=3.8.0",
    "jinja2>=3.1.3",
    "rich>=13.7.1",
    "pyarrow>=16.0.0",
]

[project.optional-dependencies]
ccxt = ["ccxt>=4.2.97"]
viz = ["plotly>=5.20.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
